# frontend/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "DMS Backend",
        "version": "0.1.0"
    },
    "paths": {
...
        "/folders": {
            "post": {
                "summary": "Create Folder",
                "operationId": "create_folder_folders_post",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/FolderCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FolderResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "get": {
                "summary": "List Folders",
                "operationId": "list_folders_folders_get",
                "parameters": [
                    {
                        "name": "parent_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "title": "Parent Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/FolderResponse"
                                    },
                                    "title": "Response List Folders Folders Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/folders/{folder_id}": {
            "get": {
                "summary": "Get Folder Contents",
                "operationId": "get_folder_contents_folders__folder_id__get",
                "parameters": [
                    {
                        "name": "folder_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Folder Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FolderContentsResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "FileResponse": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    },
                    "filename": {
                        "type": "string",
                        "title": "Filename"
                    },
                    "content_type": {
                        "type": "string",
                        "title": "Content Type"
                    },
                    "size": {
                        "type": "integer",
                        "title": "Size"
                    },
                    "category": {
                        "type": "string",
                        "title": "Category"
                    },
                    "uploaded_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Uploaded At"
                    },
                    "folder_id": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Folder Id"
                    },
                    "tags": {
                        "items": {
                            "$ref": "#/components/schemas/TagResponse"
                        },
                        "type": "array",
                        "title": "Tags",
                        "default": []
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "filename",
                    "content_type",
                    "size",
                    "category",
                    "uploaded_at",
                    "folder_id"
                ],
                "title": "FileResponse"
            },
...
            "FolderContentsResponse": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    },
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "parent_id": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Parent Id"
                    },
                    "sub_folders": {
                        "items": {
                            "$ref": "#/components/schemas/FolderResponse"
                        },
                        "type": "array",
                        "title": "Sub Folders",
                        "default": []
                    },
                    "files": {
                        "items": {
                            "$ref": "#/components/schemas/FileResponse"
                        },
                        "type": "array",
                        "title": "Files",
                        "default": []
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "name",
                    "parent_id"
                ],
                "title": "FolderContentsResponse"
            },
            "FolderCreate": {
                "properties": {
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "parent_id": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Parent Id"
                    }
                },
                "type": "object",
                "required": [
                    "name"
                ],
                "title": "FolderCreate"
            },
            "FolderResponse": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    },
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "parent_id": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Parent Id"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "name",
                    "parent_id"
                ],
                "title": "FolderResponse"
            },
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "TagResponse": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    },
                    "name": {
                        "type": "string",
                        "title": "Name"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "name"
                ],
                "title": "TagResponse"
            },
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        }
    }
}